Cambio de Long a Integer para poder usar el metodo findById de CrudRepository comento el metodo findById de la interfaz CourseRepository ya no se necesita.

+++ src/main/java/com/example/primerAppSpringBoot/repository/CourseRepository.java
@@ -3,8 +3,8 @@ package com.example.primerAppSpringBoot.repository;
 import com.example.primerAppSpringBoot.entity.Course;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CourseRepository extends CrudRepository<Course, Long> {
+public interface CourseRepository extends CrudRepository<Course, Integer> {
 
-
+    //Course findById(Integer id);
 
 }
